Update Must function return type to match Emailer interface and change template parameter type to int64

+++ email/email.go
@@ -3,5 +3,5 @@ package email
 import "context"
 
 type Emailer interface {
-	SendTemplateEmail(ctx context.Context, template int, name, email string, data map[string]any) error
+	SendTemplateEmail(ctx context.Context, template int64, name, email string, data map[string]any) error
 }
